feat: improve dashboard UI/UX with new components and route updates
- Add ErrorBoundary, LandingHero, RemediationModal, and ZombieTable components
- Update various routes and server layouts
- Enhancements to API and stores
- Fixes #43
