+++ dashboard/package.json
@@ -22,7 +22,7 @@
 		"@eslint/js": "^9.39.1",
 		"@lucide/svelte": "^0.562.0",
 		"@playwright/test": "^1.58.0",
-		"@sveltejs/adapter-node": "^5.4.0",
+		"@sveltejs/adapter-cloudflare": "^7.2.3",
 		"@sveltejs/kit": "^2.49.1",
 		"@sveltejs/vite-plugin-svelte": "^6.2.1",
 		"@tailwindcss/vite": "^4.1.18",
