+++ dashboard/src/lib/stores/jobs.svelte.ts
@@ -70,7 +70,6 @@ class JobStore {
 		});
 
 		this.#eventSource.onerror = (error) => {
-			console.error('[Jobs SSE] Error:', error);
 			this.#isConnected = false;
 			this.#eventSource?.close();
 			this.#eventSource = null;
