+++ dashboard/src/lib/tier.ts
@@ -1,6 +1,6 @@
-export type Tier = 'free_trial' | 'starter' | 'growth' | 'pro' | 'enterprise';
+export type Tier = 'free' | 'starter' | 'growth' | 'pro' | 'enterprise';
 
-const TIER_ORDER: Tier[] = ['free_trial', 'starter', 'growth', 'pro', 'enterprise'];
+const TIER_ORDER: Tier[] = ['free', 'starter', 'growth', 'pro', 'enterprise'];
 
 export function normalizeTier(value: unknown): Tier {
 	const normalized = String(value || '')
@@ -10,7 +10,7 @@ export function normalizeTier(value: unknown): Tier {
 	if (normalized === 'growth') return 'growth';
 	if (normalized === 'pro') return 'pro';
 	if (normalized === 'enterprise') return 'enterprise';
-	return 'free_trial';
+	return 'free';
 }
 
 function tierRank(tier: Tier): number {
@@ -29,11 +29,10 @@ export function tierIn(current: unknown, allowed: Tier[]): boolean {
 
 export function formatTierLabel(tier: unknown): string {
 	const normalized = normalizeTier(tier);
-	if (normalized === 'free_trial') return 'Free Trial';
+	if (normalized === 'free') return 'Free';
 	if (normalized === 'starter') return 'Starter';
 	if (normalized === 'growth') return 'Growth';
 	if (normalized === 'pro') return 'Pro';
 	if (normalized === 'enterprise') return 'Enterprise';
-	return 'Free Trial';
+	return 'Free';
 }
-
